Add android application with onCreate hook
